==> notebooks/06_plotting_lbls.ipynb <==
import pandas as pd
import sys
import numpy as np
import os
import glob
sys.path.append(os.path.join(os.path.dirname(os.getcwd()), "src"))
from overlayFunctions import Overlays
from utils import Utils

#### Plotting Single images

df_run12_meta_nf = pd.read_csv(r"Z:\__Organized_Directories_InProgress\GobyFinderDatasets\AUV_datasets\Runs\run12\run12_metadata.csv", index_col=0, low_memory=False)
df_run12_meta_nf = df_run12_meta_nf[df_run12_meta_nf.Usability == "Usable"]
filter = df_run12_meta_nf[(df_run12_meta_nf.n_fish > 10) & (df_run12_meta_nf.imh > 2176)].reset_index(drop=True)
print(len(filter))
i= 81
image = filter.image_path[i]
label = filter.label_path[i]
Overlays.disp_lbl_bbox(image, label)

path = r"D:\ageglio-1\gobyfinder_yolov8\datasets\AUV_datasets\run13\2025_goby_relabel_initial_sample"
id = "PI_1658096538_886_Iver3098_2424_0"
image = os.path.join(path, "tiled_images", id+".png")
label = os.path.join(path, "tiled_labels", id+".txt")
Overlays.disp_lbl_bbox(image, label)

#### plotting training data as a check

import pandas as pd
import sys
import numpy as np
import os
import glob
sys.path.append(os.path.join(os.path.dirname(os.getcwd()), "src"))
from overlays import Overlays

images_dir = r"..\output\inference\20240618_001_Iver3069_ABS2\tiled\images"
labels_dir = r"..\output\inference\20240618_001_Iver3069_ABS2\tiled\labels"

image_pth_file = r"Z:\__AdvancedTechnologyBackup\04_ProjectData\Innodata_2025\Goby\transects\20240804_001_Iver3069_ABS2\20240804_001_Iver3069_ABS2_batch_00\images.txt"
lbl_pth_file = r"Z:\__AdvancedTechnologyBackup\04_ProjectData\Innodata_2025\Goby\transects\20240804_001_Iver3069_ABS2\20240804_001_Iver3069_ABS2_batch_00\labels.txt"

def return_path_list(path):
    """Retrieve test images based on provided arguments."""
    try:
        print("Using list file", path)
        # --- Change made here ---
        # Use 'utf-8-sig' to automatically handle and remove the Byte Order Mark (BOM)
        with open(path, 'r', encoding='utf-8-sig') as f:
            path_list = f.read().splitlines()
        # ------------------------
        assert len(path_list) > 0, "No image paths found in csv"
    except FileNotFoundError:
        raise ValueError("Must provide an image directory or a path to a csv listing filepaths of images")
    return sorted(path_list)

def get_file_paths(args):
    """Retrieves file paths based on command-line arguments."""
    dir = os.path.abspath(args.directory)
    image_paths, label_paths = [], []
    
    # Get image paths
    if not args.labels_only:
        if args.list_file:
            image_csv = os.path.join(dir, "images.txt")
            image_paths = return_path_list(image_csv)
        else:
            image_dir = os.path.join(dir, "images")
            for ext in ['*.jpg', '*.jpeg', '*.png', '*.tif', '*.tiff']:
                image_paths.extend(glob.glob(os.path.join(image_dir, ext)))
    
    # Get label paths
    if not args.images_only:
        if args.list_file:
            label_csv = os.path.join(dir, "labels.txt")
            label_paths = return_path_list(label_csv)
        else:
            label_dir = os.path.join(dir, "labels")
            label_paths.extend(glob.glob(os.path.join(label_dir, '*.txt')))
    assert len(image_paths) > 0 or len(label_paths) > 0, "No image or label files found in the specified directory."
    assert len(image_paths) == len(label_paths) or args.labels_only or args.images_only, \
        "Mismatch between image and label counts. Ensure both directories have matching files or use --images_only or --labels_only."
    return sorted(image_paths), sorted(label_paths)

def main():
    args = parse_arguments()
    
    # Enforce dimension arguments in labels_only mode
    if args.labels_only and (args.image_width is None or args.image_height is None):
        raise ValueError("Error: --image_width and --image_height are required in --labels_only mode.")

    image_paths, label_paths = get_file_paths(args)
    plot_img =  Overlays.disp_lbl_bbox(image, label)
    plot_img.show()

# 321 446 538 1206 1320 1891 2080 2783 3186 3781 4349 4456 5118 5473 6239 - alewife 6412 7410 7887 8292 8364 8665 8846 9085 9958 10508 11178 11429 11436 11863 12445 12502 12579 12983 13096 13160
import random
i = random.randint(0, len(matched_images))
# i = random.randint(0, 300)
print(i)
image = matched_images[i]
label = matched_labels[i]
plot_img =  Overlays.disp_lbl_bbox(image, label)
plot_img

#### plotting and saving OBB QAQC images for alewife detections

# img_pth = r"Z:\__Organized_Directories_InProgress\Annotated_Images\2021_LakeMichigan_Annotated_Images\images\CI_1625690871_595.png"
years = ['2020', '2021', '2022', '2023']
years = ["2022"]
def plot_alewife_images(years):
    for year in years:
        img_dir = f"datasets\\AUV_datasets\\images\\{year}"
        run_dir = f"test_runs\\detect\\Run14+-alewife_annot_imgs_{year}"
        lbl_csv = f"Run14+-alewife_annot_imgs_{year}_Labels.csv"
        score_csv = f"Run14+-alewife_annot_imgs_{year}_scores.csv"
        lbl_df = pd.read_csv(os.path.join(run_dir, lbl_csv))
        scores_df = pd.read_csv(os.path.join(run_dir, score_csv))
        new_path = os.path.join(run_dir, f"{year}_qaqc_images")
        if not os.path.exists(new_path): os.mkdir(new_path)
        for im in scores_df.image_id.unique():
            im_pth = glob.glob(os.path.join(img_dir, im+".jpg"))[0]
            Overlays().save_annot_imgs_obb_2(im_pth, lbl_df, scores_df, conf_thresh=0.0, new_path=new_path)

### Generate prediction labels only for LMGS Gopro QAQC

# import pandas as pd
# img_pths = general.list_files("datasets\LMBS Goby Photos", ".jpg", gopro=True)
# new_path = r"C:\Users\ageglio\ageglio-1\gobyfinder_yolov8\inference\LMBS_Goby_Photos_conf0.05\lbl_imgs"
# pred_df = pd.read_csv(r"C:\Users\ageglio\ageglio-1\gobyfinder_yolov8\inference\LMBS_Goby_Photos_conf0.05\LMBS_Goby_Photos_conf0.05_Yolo_predictions.csv")
# for img_pth in img_pths:
#     generate_annot_imgs().save_annot_imgs_pred_only(img_pth, pred_df, new_path, background = "image")

### Generate annotated images with labels and predictions

score_df = pd.read_csv(r"test_runs\detect\balanced_test_init_assmt\balanced_test_init_assmt_scores.csv")
simid = sorted(score_df.Filename.unique())
img_pths = [r"Z:\__Organized_Directories_InProgress\Fish_annotations_based_on_YOLO\2020\screened_with_fish\original_images\20200809_001_Iver3069_ABS1\PI_1596974969_825_Iver3069.png"]
imid = sorted(list(map(lambda x: os.path.basename(x).split(".")[0], img_pths)))
new_path = r"test_runs\detect\balanced_test_init_assmt"
assert set(imid).intersection(simid) == set(imid), "img_pths not contained in validation labels"
# for img_pth in img_pths:
#     generate_annot_imgs().save_annot_imgs(img_pth, score_df, new_path, conf_thresh=0.36)



names = ['PI_1596974969_825_Iver3069_dt_107','PI_1596974969_825_Iver3069_dt_111','PI_1596974969_825_Iver3069_dt_113','PI_1596974969_825_Iver3069_dt_118','PI_1596974969_825_Iver3069_dt_116','PI_1596974969_825_Iver3069_dt_117','PI_1596974969_825_Iver3069_dt_119',        'PI_1596974969_825_Iver3069_dt_114',        'PI_1596974969_825_Iver3069_dt_101',        'PI_1596974969_825_Iver3069_dt_90',        'PI_1596974969_825_Iver3069_dt_121',        'PI_1596974969_825_Iver3069_dt_72',        'PI_1596974969_825_Iver3069_dt_63',        'PI_1596974969_825_Iver3069_dt_98']
names_small = list(map(lambda x: x.split("_")[-1], names))
wts = [0.517554073,0.755342822,0.557671634,0.256169483,0.392054384,0.496877683,0.46887362,0.496877683,0.632590853,0.480832252,0.513620776,0.187196292,0.288527773,0.405346274]
df = pd.DataFrame(np.c_[names_small, wts], columns=["lbl_ID", "weight (g)"])
df["weight (g)"] = df["weight (g)"].astype(float).round(3)

### Generate labels for Inference QAQC

from src import generate_labels
import os
import pandas as pd
# def generate_labels_for_inference_QAQC(confidence, img_pths, preds, new_path):
#     preds_imgs = preds.Filename.unique()
#     filenames = img_pths.apply(lambda x: os.path.basename(x).split(".")[0])
#     print(len(filenames))
#     filenamesp = filenames[filenames.isin(preds_imgs)]    
#     img_pths = img_pths[filenames.isin(preds_imgs)]
#     print(len(img_pths))
#     # img_pths_sample = img_pths.sample(100, random_state=42)
#     pred_df = preds[preds.Filename.isin(filenamesp)]
#     assert pd.Series(pred_df.Filename.unique()).isin(filenames).all() == True
#     if not os.path.exists(new_path):
#         os.makedirs(new_path)
#     i = 1
#     for img_pth in img_pths:
#         l = len(img_pths)
#         if not os.path.exists(os.path.join(new_path, "anno_"+os.path.basename(img_pth))):
#             print("generating", i,"/",l, end="   \r")
#             generate_annot_imgs().save_annot_imgs_pred_only(confidence, img_pth, pred_df, new_path, background = "image")
#         i+=1
#     print("\n complete")





### Mussel Control 2024



def generate_labels_for_inference_QAQC(conf_thresh, img_pths, pred_df, new_path, starting_i):
    i = 1
    for img_pth in img_pths:
        l = len(img_pths)
        print("generating", i,"/",l, end="   \r")
        if i <= starting_i:
            i += 1
            return
        else:
            generate_annot_imgs.save_annot_imgs_pred_only(img_pth, pred_df, new_path, conf_thresh, background = "image")
            i += 1
    print("\n complete")

pred_df = pd.read_csv(r"Z:\Proj_Yolo\Mussel_Control_2024\Mussel_control_GobyFinder_predictions.csv", index_col=0)
img_pths = glob.glob(r"G:\mussel_control_2024\*.png")
new_path = r"Z:\Proj_Yolo\Mussel_Control_2024\labeled_images"
conf_thresh = 0.6
generate_labels_for_inference_QAQC(conf_thresh, img_pths, pred_df, new_path)

### 2020 inference qaqc

confidence = 0.28
assmt_folder = r"Z:\Proj_Yolo\2020_Yolo_update\02_2020_Assessment\2020_Assessment_11-2"
preds = pd.read_csv(os.path.join(assmt_folder, "2020_Yolo_predictions.csv"))
img_pths = pd.read_csv(os.path.join(assmt_folder, "meta_2020_usable_goby.csv"))["image_path"]
new_path = r"Z:\Proj_Yolo\Assessment_11-2_imgs\2020_lbl_imgs"
# samp = generate_labels_for_inference_QAQC(confidence, img_pths, preds)

### 2021 inference

confidence = 0.13
assmt_folder = r"Z:\Proj_Yolo\2021_Yolo_update\02_2021_Assessment\2021_Assessment_11-2"
img_pths = pd.read_csv(os.path.join(assmt_folder, "meta_2021_usable_goby.csv"))["image_path"]
preds = pd.read_csv(os.path.join(assmt_folder, "2021_Yolo_predictions.csv"))
new_path = r"Z:\Proj_Yolo\Assessment_11-2_imgs\2021_lbl_imgs"
# generate_labels_for_inference_QAQC(confidence, img_pths, preds)

#### 2022

confidence = 0.11
assmt_folder = r"Z:\Proj_Yolo\2022_Yolo_update\02_2022_Assessment\2022_Assessment_11-2"
img_pths = pd.read_csv(os.path.join(assmt_folder, "meta_2022_usable_goby.csv"))["image_path"]
preds = pd.read_csv(os.path.join(assmt_folder, "2022_Yolo_predictions.csv"))
new_path = r"Z:\Proj_Yolo\Assessment_11-2_imgs\2022_lbl_imgs"
# generate_labels_for_inference_QAQC(confidence, img_pths, preds)

#### 2023

confidence = 0.10
assmt_folder = r"Z:\Proj_Yolo\2023_Yolo_update\02_2023_Assessment\2023_Assessment_11-2"
img_pths = pd.read_csv(os.path.join(assmt_folder, "meta_2023_usable_goby.csv"))["image_path"]
preds = pd.read_csv(os.path.join(assmt_folder, "2023_Yolo_predictions.csv"))
new_path = r"Z:\Proj_Yolo\Assessment_11-2_imgs\2023_lbl_imgs"
# generate_labels_for_inference_QAQC(confidence, img_pths, preds)
